pyproject.toml
[project]
name = "insane-whisper-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.2.1",
    "bcrypt>=4.2.1",
    "fastapi>=0.115.6",
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "uvicorn>=0.34.0",
]
